Obstacles/bonus/sound
I added the image of the obstacles which were birds, boxes, and fences.
Then, also added a sound effect when the bunny jump and when the game is over.
For the carrots, I had added image of a carrot which is indicating a bonus points. So I made the carrot moved setting its velocity to make it move over time and the speed increases over time. However, when you collide with the carrot as much as the time you are colliding you get extra points by 0.5